+++ src/PimcoreNavigationBundle.php
@@ -1,10 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GalDigitalGmbh\PimcoreNavigation;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class PimcoreNavigationBundle extends Bundle
+final class PimcoreNavigationBundle extends Bundle
 {
-
 }